feat(react-router): added react-router-dom and implemented de route /personagens to the page
